requirements.txt
aiohttp
asyncio
faiss-cpu
fastapi[standard]
numpy
pandas
torch
torchvision
transformers
webdataset
albumentations==1.4.16
img2dataset==1.44.1